Extract Groovy/Kotlin DSL specifics in dedicated types
